fix: switch dagSize type to Long (#116)

+++ packages/cron/src/jobs/pins.js
@@ -33,7 +33,7 @@ const CREATE_OR_UPDATE_PIN = gql`
 `
 
 const UPDATE_CONTENT_DAG_SIZE = gql`
-  mutation UpdateContentDagSize($content: ID!, $dagSize: Int!) {
+  mutation UpdateContentDagSize($content: ID!, $dagSize: Long!) {
     updateContentDagSize(content: $content, dagSize: $dagSize) {
       _id
     }
